Added basic unittesting for GitHubPackage
Moved a few items around and added unittesting.
