+++ autoupgrade/__init__.py
@@ -1,17 +1,9 @@
 # -*- coding: utf-8 -*-
 
-from .exceptions import NoVersionsError, PIPError, PkgNotFoundError
-from .package import Package
-from .utils import normalize_version
 
+from .package import (Package, AutoUpgrade)
 
-# NOTE: Legacy class
-class AutoUpgrade(Package):
-
-    def upgrade(self, *args, **kwargs):
-        try:
-            Package.upgrade(self, *args, **kwargs)
-        except PIPError:
-            return False
-        else:
-            return True
+__all__ = [
+    "Package",
+    "AutoUpgrade",
+]
