+++ tests/test_autoupgrade.py
@@ -2,7 +2,8 @@
 
 from unittest import TestCase
 
-from autoupgrade import Package, ver_to_tuple
+from autoupgrade import Package
+from autoupgrade.utils import ver_to_tuple
 
 
 class TestFunctions(TestCase):
@@ -27,12 +28,23 @@ class TestFunctions(TestCase):
             ver_to_tuple('1.2.3'),
             ver_to_tuple('1.2.3'))
 
-    def test_upgrade_default(self):
-        inst = Package("pip", verbose=True)
+    def test_pypi_upgrade_default(self):
+        inst = Package.pypi(
+            pkg="pip",
+            verbose=True)
         inst.smartupgrade(restart=False)
 
-    def test_upgrade_index(self):
-        inst = Package("pip",
-                       "https://pypi.python.org/simple",
-                       verbose=True)
+    def test_pypi_upgrade_index(self):
+        inst = Package.pypi(
+            pkg="pip",
+            index="https://pypi.python.org/simple",
+            verbose=True)
+        inst.smartupgrade(restart=False)
+
+    def test_github_upgrade_default(self):
+        inst = Package.github(
+            pkg="autoupgrade",
+            user="vuolter",
+            repo="autoupgrade-ng",
+            verbose=True)
         inst.smartupgrade(restart=False)
